Refactor ChatInput component to improve structure and readability; implement new subcomponents for file upload and message actions, enhancing maintainability and user experience.
